Refactor app into modular backend and frontend
